Update the folder structure
