+++ src/pages/_app.tsx
@@ -1,29 +1,19 @@
 import type {AppProps} from 'next/app';
 import React from 'react';
 import PropTypes from 'prop-types';
-import Head from 'next/head';
 import { ThemeProvider } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import theme from '../utils/theme';
 
 const MyApp = ({Component, pageProps}: AppProps) => {
-  React.useEffect(() => {
-    // Remove the server-side injected CSS.
-    const jssStyles = document.querySelector('#jss-server-side');
-    if (jssStyles) {
-      jssStyles.parentElement.removeChild(jssStyles);
-    }
-  }, []);
-  return (
-    <React.Fragment>
-      <Head>
-      </Head>
-      <ThemeProvider theme={theme}>
-        <CssBaseline />
-        <Component {...pageProps} />
-      </ThemeProvider>
-    </React.Fragment>
-  );
+    return (
+        <React.Fragment>
+            <ThemeProvider theme={theme}>
+                <CssBaseline />
+                <Component {...pageProps} />
+            </ThemeProvider>
+        </React.Fragment>
+    );
 };
 
 // Only uncomment this method if you have blocking data requirements for
@@ -38,8 +28,8 @@ const MyApp = ({Component, pageProps}: AppProps) => {
 //   return { ...appProps }
 // }
 MyApp.propTypes = {
-  Component: PropTypes.elementType.isRequired,
-  pageProps: PropTypes.object.isRequired,
+    Component: PropTypes.elementType.isRequired,
+    pageProps: PropTypes.object.isRequired,
 };
 
 export default MyApp;
